[mlir] Ptr dialect (1/n)
This is the first patch of a series introducing the pointer dialect, a dialect
to model pointer operations motivated by the goal of modularizing the LLVM
dialect.

This patch creates the pointer dialect, extracts the LLVM Pointer type from LLVM
and generalize it to accept any attribute as a memory space, while also
maintining the stability of the tree and tests.