[mlir][CMake] This patch modifies the behavior of `add_mlir_aggregate` and adds `mlir_c_target_link_libraries`.
This patch modifies `add_mlir_aggregate` to link against `libMLIR.so` if
`MLIR_LINK_MLIR_DYLIB` is set. This means that aggregates that are present in
`libMLIR.so` are not added to the final aggregate library.

This allows using `libMLIR.so` as a single source of truth, making MLIR
easier to be distributed.

This patch also adds `mlir_c_target_link_libraries`, the behavior is identical
to `mlir_target_link_libraries` except it targets `libMLIR-C.so`.

It also makes the python CAPI depend on `libMLIR-C.so` if
`MLIR_BUILD_MLIR_C_DYLIB=1`.

Finally, it fixes some ill-configured library dependencies.